feat: implement managed cluster data source and related models
