+++ internal/provider/managed-cluster/managed_cluster_data_source.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package provider
+package managed_cluster
 
 import (
 	"context"
@@ -23,54 +23,6 @@ var (
 
 type ManagedClusterDataSource struct {
 	client *api_v2025.APIClient
-}
-
-type ManagedClusterDataSourceModel struct {
-	ID                   types.String             `tfsdk:"id"`
-	Name                 types.String             `tfsdk:"name"`
-	Pod                  types.String             `tfsdk:"pod"`
-	Org                  types.String             `tfsdk:"org"`
-	Type                 types.String             `tfsdk:"type"`
-	Configuration        types.Map                `tfsdk:"configuration"`
-	KeyPair              ManagedClusterKeyPair    `tfsdk:"key_pair"`
-	Attributes           ManagedClusterAttributes `tfsdk:"attributes"`
-	Description          types.String             `tfsdk:"description"`
-	Redis                ManagedClusterRedis      `tfsdk:"redis"`
-	ClientType           types.String             `tfsdk:"client_type"`
-	CcgVersion           types.String             `tfsdk:"ccg_version"`
-	PinnedConfig         types.Bool               `tfsdk:"pinned_config"`
-	Operational          types.Bool               `tfsdk:"operational"`
-	Status               types.String             `tfsdk:"status"`
-	PublicKeyCertificate types.String             `tfsdk:"public_key_certificate"`
-	PublicKeyThumbprint  types.String             `tfsdk:"public_key_thumbprint"`
-	PublicKey            types.String             `tfsdk:"public_key_type"`
-	AlertKey             types.String             `tfsdk:"alert_key"`
-	ClientIds            types.List               `tfsdk:"client_ids"`
-	ServiceCount         types.Int32              `tfsdk:"service_count"`
-	CcId                 types.String             `tfsdk:"cc_id"`
-	CreatedAt            types.String             `tfsdk:"created_at"`
-	UpdatedAt            types.String             `tfsdk:"updated_at"`
-}
-
-type ManagedClusterKeyPair struct {
-	PublicKey            types.String `tfsdk:"public_key"`
-	PublicKeyThumbprint  types.String `tfsdk:"public_key_thumbprint"`
-	PublicKeyCertificate types.String `tfsdk:"public_key_certificate"`
-}
-
-type ManagedClusterAttributes struct {
-	Queue    ManagedClusterQueue `tfsdk:"queue"`
-	Keystore types.String        `tfsdk:"keystore"`
-}
-
-type ManagedClusterQueue struct {
-	Name   types.String `tfsdk:"name"`
-	Region types.String `tfsdk:"region"`
-}
-
-type ManagedClusterRedis struct {
-	RedisHost types.String `tfsdk:"redis_host"`
-	RedisPort types.Int32  `tfsdk:"redis_port"`
 }
 
 func NewManagedClusterDataSource() datasource.DataSource {
